Split a big file into smaller files in http mixer
