+++ src/envoy/http/mixer/BUILD
@@ -23,9 +23,18 @@ load(
 envoy_cc_library(
     name = "filter_lib",
     srcs = [
+        "check_data.cc",
+        "check_data.h",
+        "config.cc",
         "config.h",
-        "http_filter.cc",
-        "mixer_control.h",
+        "control.cc",
+        "control.h",
+        "control_factory.h",
+        "filter.cc",
+        "filter.h",
+        "filter_factory.cc",
+        "header_update.h",
+        "report_data.h",
     ],
     repository = "@envoy",
     visibility = ["//visibility:public"],
